Nighthawk-PoC: state dump - split out benchmark_loop.cc/h - fix up includes in client.cc/h
