+++ source/exe/BUILD
@@ -46,12 +46,14 @@ envoy_cc_library(
         "stream_decoder.cc",
 	    "client_options_impl.cc",
 	    "ssl.cc",
+        "benchmark_loop.cc",
     ],
     hdrs = [
         "client.h",
         "stream_decoder.h",
     	"client_options_impl.h",
         "ssl.h",
+        "benchmark_loop.h",
     ],
     repository = "@envoy",
     deps = [
